+++ src/models/Profile.ts
@@ -1,5 +1,11 @@
-export type Reference = {
+export interface Tagset {
   name: string;
-  uri?: string;
+  tags: Array<string>;
+}
+
+export interface Reference {
+  id: string;
+  name: string;
+  uri: string;
   description?: string;
-};
+}
